+++ packages/orchestratorlib/src/datasourcehelper.ts
@@ -128,8 +128,7 @@ export class DataSourceHelper {
         }
         input.FilePath = path.join(filePath, input.RoutingName + '.lu');
       } else if (input.Type === 'qna') {
-          input.RoutingName = Utility.isEmptyString(input.RoutingName) ? `q_${input.Id}` : `${input.RoutingName}`;
-        }
+        input.RoutingName = Utility.isEmptyString(input.RoutingName) ? `q_${input.Id}` : `${input.RoutingName}`;
         input.FilePath = path.join(filePath, input.RoutingName + '.qna');
       } else {
         throw new Error(`Invalid content for type ${input.Type}`);
