+++ packages/orchestratorlib/src/orchestratorhelper.ts
@@ -249,6 +249,24 @@ export class OrchestratorHelper {
     return entitiesList;
   }
 
+  public static jsonStringify(obj: any): string {
+    return JSON.stringify(
+      obj,
+      (key: string, value: any) => {
+        if (value && typeof value === 'object' && !Array.isArray(value)) {
+          const replacement: any = {};
+          for (const k in value) {
+            if (Object.hasOwnProperty.call(value, k)) {
+              replacement[k && k.charAt(0).toLowerCase() + k.substring(1)] = value[k];
+            }
+          }
+          return replacement;
+        }
+        return value;
+      },
+      2);
+  }
+
   // eslint-disable-next-line max-params
   static async processFile(
     filePath: string,
